Docs: Add docs for test-e2e script added to scripts packages
