+++ packages/scripts/README.md
@@ -20,8 +20,11 @@ _Example:_
 {
 	"scripts": {
 		"check-engines": "wp-scripts check-engines",
+		"check-licenses": "wp-scripts check-licenses --production",
+		"lint:js": "wp-scripts lint-js .",
 		"lint:pkg-json": "wp-scripts lint-pkg-json .",
-		"test": "wp-scripts test-unit-js"
+		"test:e2e": "wp-scripts test-e2e",
+		"test:unit": "wp-scripts test-unit-js"
 	}
 }
 ```
@@ -45,8 +48,28 @@ _Example:_
 This is how you execute the script with presented setup:
 * `npm run check-engines` - checks installed version of `node` and `npm`.
 
+### `check-licenses`
 
-### `wp-scripts lint-js`
+Validates that all dependencies of a project are compatible with the project's own license.
+
+_Example:_
+
+```json
+{
+	"scripts": {
+		"check-licenses": "wp-scripts check-licenses --prod --gpl2 --ignore=abab"
+	}
+}
+```
+
+_Flags_:
+
+- `--prod` (or `--production`): When present, validates only `dependencies` and not `devDependencies`
+- `--dev` (or `--development`): When present, validates both `dependencies` and `devDependencies`
+- `--gpl2`: Validates against [GPLv2 license compatibility](https://www.gnu.org/licenses/license-list.en.html)
+- `--ignore=a,b,c`: A comma-separated set of package names to ignore for validation. This is intended to be used primarily in cases where a dependency's `license` field is malformed. It's assumed that any `ignored` package argument would be manually vetted for compatibility by the project owner.
+
+### `lint-js`
 
 Helps enforce coding style guidelines for your JavaScript files. It uses [eslint](https://eslint.org/) with the set of recommended rules defined in [@wordpress/eslint-plugin](https://www.npmjs.com/package/@wordpress/eslint-plugin) npm package. You can override default rules with your own as described in [eslint docs](https://eslint.org/docs/rules/).
 
@@ -63,7 +86,7 @@ _Example:_
 This is how you execute the script with presented setup:
 * `npm run lint:js` - lints JavaScripts files in the whole project's.
 
-### `wp-scripts lint-pkg-json`
+### `lint-pkg-json`
 
 Helps enforce standards for your package.json files. It uses [npm-package-json-lint](https://www.npmjs.com/package/npm-package-json-lint) with the set of recommended rules defined in [@wordpress/npm-package-json-lint-config](https://www.npmjs.com/package/@wordpress/npm-package-json-lint-config) npm package. You can override default rules with your own as described in [npm-package-json-lint wiki](https://github.com/tclindner/npm-package-json-lint/wiki).
 
@@ -80,52 +103,59 @@ _Example:_
 This is how you execute those scripts using the presented setup:
 * `npm run lint:pkg-json` - lints `package.json` file in the project's root folder.
 
-### `wp-scripts test-unit-js`
+### `test-e2e`
 
-_Alias_: `wp-scripts test-unit-jest` 
+Launches the End-To-End (E2E) test runner. It uses [Jest](https://facebook.github.io/jest/) behind the scenes and you are able to utilize all of its [CLI options](https://facebook.github.io/jest/docs/en/cli.html). You can also run `./node_modules/.bin/wp-scripts test:e2e --help` or `npm run test:e2e:help` (as presented below) to view all of the available options.
 
-Launches the test runner. It uses [Jest](https://facebook.github.io/jest/) behind the scenes and you are able to utilize all of its [CLI options](https://facebook.github.io/jest/docs/en/cli.html). You can also run `./node_modules/.bin/wp-scripts test-unit-js --help` or `npm run test:help` (as presented below) to view all of the available options. By default, it uses the set of recommended options defined in [@wordpress/jest-preset-default](https://www.npmjs.com/package/@wordpress/jest-preset-default) npm package. You can override them with your own options as described in [Jest documentation](https://jestjs.io/docs/en/configuration).
+Writing tests can be done using Puppeteer API:
+ 
+> [Puppeteer](https://pptr.dev/) is a Node library which provides a high-level API to control Chrome or Chromium over the [DevTools Protocol](https://chromedevtools.github.io/devtools-protocol/). Puppeteer runs [headless](https://developers.google.com/web/updates/2017/04/headless-chrome) by default, but can be configured to run full (non-headless) Chrome or Chromium.
 
 _Example:_
 
 ```json
 {
 	"scripts": {
-		"test": "wp-scripts test-unit-js",
-		"test:help": "wp-scripts test-unit-js --help",
-		"test:watch": "wp-scripts test-unit-js --watch"
+		"test:e2e": "wp-scripts test-e2e",
+		"test:e2e:help": "wp-scripts test-e2e --help"
 	}
 }
 ```
 
 This is how you execute those scripts using the presented setup:
-* `npm run test` or `npm test` - runs all unit tests.
-* `npm run test:help` - prints all available options to configure unit tests runner.
-* `npm run test:watch` - runs all unit tests in the watch mode.
+* `npm run test:e2e` - runs all unit tests.
+* `npm run test:e2e:help` - prints all available options to configure unit tests runner.
 
-### `wp-scripts check-licenses`
+This script automatically detects the best config to start Puppeteer but sometimes you may need to specify custom options:
+ - You can add a `jest-puppeteer.config.js` at the root of the project or define a custom path using `JEST_PUPPETEER_CONFIG` environment variable. Check [jest-puppeteer](https://github.com/smooth-code/jest-puppeteer#jest-puppeteerconfigjs) for more details.
 
-Validates that all dependencies of a project are compatible with the project's own license.
+We enforce that all tests run serially in the current process using [--runInBand](https://jestjs.io/docs/en/cli#runinband) Jest CLI option to avoid conflicts between tests caused by the fact that they share the same WordPress instance.
+
+### `test-unit-js`
+
+_Alias_: `test-unit-jest` 
+
+Launches the unit test runner. It uses [Jest](https://facebook.github.io/jest/) behind the scenes and you are able to utilize all of its [CLI options](https://facebook.github.io/jest/docs/en/cli.html). You can also run `./node_modules/.bin/wp-scripts test-unit-js --help` or `npm run test:unit:help` (as presented below) to view all of the available options. By default, it uses the set of recommended options defined in [@wordpress/jest-preset-default](https://www.npmjs.com/package/@wordpress/jest-preset-default) npm package. You can override them with your own options as described in [Jest documentation](https://jestjs.io/docs/en/configuration).
 
 _Example:_
 
 ```json
 {
 	"scripts": {
-		"check-licenses": "wp-scripts check-licenses --prod --gpl2 --ignore=abab",
+		"test:unit": "wp-scripts test-unit-js",
+		"test:unit:help": "wp-scripts test-unit-js --help",
+		"test:unit:watch": "wp-scripts test-unit-js --watch"
 	}
 }
 ```
 
-_Flags_:
-
-- `--prod` (or `--production`): When present, validates only `dependencies` and not `devDependencies`
-- `--dev` (or `--development`): When present, validates both `dependencies` and `devDependencies`
-- `--gpl2`: Validates against [GPLv2 license compatibility](https://www.gnu.org/licenses/license-list.en.html)
-- `--ignore=a,b,c`: A comma-separated set of package names to ignore for validation. This is intended to be used primarily in cases where a dependency's `license` field is malformed. It's assumed that any `ignored` package argument would be manually vetted for compatibility by the project owner.
+This is how you execute those scripts using the presented setup:
+* `npm run test:unit` - runs all unit tests.
+* `npm run test:unit:help` - prints all available options to configure unit tests runner.
+* `npm run test:unit:watch` - runs all unit tests in the watch mode.
 
 ## Inspiration
 
-This is inspired by [react-scripts](https://www.npmjs.com/package/react-scripts) and [kcd-scripts](https://www.npmjs.com/package/kcd-scripts).
+This package is inspired by [react-scripts](https://www.npmjs.com/package/react-scripts) and [kcd-scripts](https://www.npmjs.com/package/kcd-scripts).
 
 <br/><br/><p align="center"><img src="https://s.w.org/style/images/codeispoetry.png?1" alt="Code is Poetry." /></p>
